feat: funcionalidades nas CRUDs de usuario e review

+++ backend/controllers/reviewsC.js
@@ -4,9 +4,23 @@ import pool from '../db.js';
 // CREATE
 export const criarReview = async (req, res) => {
     try {
-        const { nota, texto, idfilme, idusuario } = req.body;
+        const { idfilme, idusuario } = req.params; 
+        const { nota, texto } = req.body;
+
+        // Verifica se o filme existe
+        const filme = await pool.query('SELECT * FROM filmes WHERE id = $1', [idfilme]);
+        if (filme.rows.length === 0) {
+            return res.status(404).json({ erro: 'Filme não encontrado' });
+        }
+
+        // Verifica se o usuário existe
+        const usuario = await pool.query('SELECT * FROM usuarios WHERE id = $1', [idusuario]);
+        if (usuario.rows.length === 0) {
+            return res.status(404).json({ erro: 'Usuário não encontrado' });
+        }
+
         const novaReview = await pool.query(
-            'INSERT INTO filmes (nota, texto, idfilme, idusuario) VALUES ($1, $2, $3, $4) RETURNING *',
+            'INSERT INTO reviews (nota, texto, idfilme, idusuario) VALUES ($1, $2, $3, $4) RETURNING *',
             [nota, texto , idfilme, idusuario]
         );
         res.json(novaReview.rows[0]);
@@ -19,13 +33,15 @@ export const criarReview = async (req, res) => {
 // READ
 export const obterReviewPorUsuario = async (req, res) => {
     try {
-        const { idUsuario } = req.params;
-        const todasReviewsUser = await pool.query('SELECT * FROM reviews WHERE idusuario = $1', [idUsuario]);
-        res.json(todasReviewsUser.rows);
+        const { idusuario } = req.params;
+        const todasReviewsUser = await pool.query('SELECT * FROM reviews WHERE idusuario = $1', [idusuario]);
 
-        if (todasReviewsUser=== 0) {
+        if (todasReviewsUser.rows.length === 0) {
             return res.status(404).json({ erro: 'Usuário sem reviews!' });
+
         }
+
+        res.json(todasReviewsUser.rows);        
 
     } catch (err) {
         console.error(err.message);
@@ -35,13 +51,16 @@ export const obterReviewPorUsuario = async (req, res) => {
 
 export const obterReviewPorFilme = async (req, res) => {
     try {
-        const { idFilme } = req.params;
-        const todasReviewsFilme = await pool.query('SELECT * FROM reviews WHERE id = $1', [idFilme]);
-        res.json(todasReviewsFilme.rows[0]);
+        const { idfilme } = req.params;
+        const todasReviewsFilme = await pool.query('SELECT * FROM reviews WHERE idfilme = $1', [idfilme]);
 
-        if (todasReviewsFilme=== 0) {
+        if (todasReviewsFilme.rows.length === 0) {
             return res.status(404).json({ erro: 'Filme sem reviews!' });
         }
+
+        res.json(todasReviewsFilme.rows);
+
+        
 
     } catch (err) {
         console.error(err.message);
@@ -53,7 +72,7 @@ export const obterReviewFilmePorUsuario = async (req, res) => {
     try {
         const { idFilme, idUsuario } = req.params;
         const reviewFilmePorUsuario = await pool.query('SELECT * FROM reviews WHERE idfilme = $1 AND idusuario = $2', [idFilme, idUsuario]);
-        res.json(reviewFilmePorUsuario.rows[0]);
+        res.json(reviewFilmePorUsuario.rows);
 
         if (reviewFilmePorUsuario.rows.length === 0) {
             return res.status(404).json({ erro: 'Review não encontrada' });
@@ -74,6 +93,12 @@ export const atualizarReview = async (req, res) => {
         const camposParaAtualizar = [];
         const valores = [];
 
+        // Verifica se a review existe
+        const review = await pool.query('SELECT * FROM reviews WHERE id = $1', [id]);
+        if (review.rows.length === 0) {
+            return res.status(404).json({ erro: 'Review não encontrada' });
+        }
+        
          if (nota !== undefined) {
             camposParaAtualizar.push('nota = $' + (valores.length + 1));
             valores.push(nota);
@@ -109,8 +134,16 @@ export const atualizarReview = async (req, res) => {
 export const deletarReview = async (req, res) => {
     try {
         const { id } = req.params;
+
+        // Verifica se a review existe
+        const review = await pool.query('SELECT * FROM reviews WHERE id = $1', [id]);
+        if (review.rows.length === 0) {
+            return res.status(404).json({ erro: 'Review não encontrada' });
+        }
+
         await pool.query('DELETE FROM reviews WHERE id = $1', [id]);
         res.json('Review deletada com sucesso!');
+
     } catch (err) {
         console.error(err.message);
         res.status(500).send('Erro no servidor');
